Add streaming API and utility functions to Rust and Go implementations
Expands Rust and Go implementations to feature parity with JS/TS:

Streaming API:
- SlimEncoder: incremental encoding with buffering
- SlimDecoder: incremental decoding with chunk accumulation
- encode_chunked/EncodeChunked: split large datasets into chunks
- decode_iter/DecodeIter: iterate over decoded objects
- DecodeStreamChan (Go): channel-based streaming for concurrency
- Collect: gather async/channel results

Utility Functions:
- deep_equal/DeepEqual: deep comparison with NaN handling
- clone/Clone: deep copy of values
- get_path/GetPath: dot-notation path access (e.g., "user.name")
- set_path/SetPath: immutable path-based updates
- estimate_tokens/EstimateTokens: LLM token estimation
- calculate_savings/CalculateSavings: SLIM vs JSON efficiency

Testing:
- Added property-based tests using proptest (Rust) and gopter (Go)
- Tests verify roundtrip consistency, encode stability, and API contracts
- Cross-implementation compatibility tests for consistent behavior

Documentation:
- Updated README with complete API reference for all implementations
- Added comprehensive examples for streaming and utility functions

+++ rust/src/utils.rs
@@ -38,15 +38,15 @@ pub fn estimate_tokens(s: &str) -> usize {
 ///
 /// let savings = calculate_savings("|2|id#|\\n1\\n2", "[{\"id\":1},{\"id\":2}]");
 /// ```
-pub fn calculate_savings(slim_str: &str, json_str: &str) -> u32 {
-    let slim_tokens = estimate_tokens(slim_str);
-    let json_tokens = estimate_tokens(json_str);
+pub fn calculate_savings(slim_str: &str, json_str: &str) -> i32 {
+    let slim_tokens = estimate_tokens(slim_str) as i32;
+    let json_tokens = estimate_tokens(json_str) as i32;
 
     if json_tokens == 0 {
         return 0;
     }
 
-    (((json_tokens - slim_tokens) as f64 / json_tokens as f64) * 100.0).round() as u32
+    (((json_tokens - slim_tokens) as f64 / json_tokens as f64) * 100.0).round() as i32
 }
 
 /// Deep equality check for JSON values (handles NaN correctly).
